requirements.txt
accelerate==0.32.1
datasets==2.20.0
diffusers==0.29.2
dvc==3.51.2
hydra-core==1.3.2
omegaconf==2.3.0
packaging==24.1
peft==0.11.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.3
wandb==0.17.4
fastapi==0.111.0
pillow==10.4.0
pyyaml==6.0.1
requests==2.32.3
